requirements.txt
Flask==3.1.0
flask-mysql-connector==1.1.0
mysql-connector-python==9.2.0
python-dotenv==1.1.0
openpyxl==3.1.5
waitress==3.0.2
pypdf>=4.0.0
reportlab>=4.0.0
Pillow>=10.0.0
